spv: prepare for multiple ext inst sets
