+++ naga/src/back/spv/index.rs
@@ -200,7 +200,7 @@ impl<'w> BlockContext<'w> {
         // BoundsCheckPolicy::Restrict.
         let restricted_index_id = self.gen_id();
         block.body.push(Instruction::ext_inst(
-            self.writer.gl450_ext_inst_id,
+            self.writer.extension_inst_import("GLSL.std.450"),
             spirv::GLOp::UMin,
             self.writer.get_uint_type_id(),
             restricted_index_id,
